Refine data collection and sorting in transaction processing scripts
This commit modifies the output messages in prep_monthly_transaction_xero_csv.py for better clarity on collected data and time periods. Additionally, it refines the sequence of data processing in data_processing_tools.py, by implementing sorted processing of JSON files and adjusting the assignment of the '*Date' attribute in the data row. This improves the positional and temporal precision of transactions recorded.